+++ app/lib/ExtraLib.h
@@ -1,0 +1,7 @@
+#ifndef ExtraLib
+#define ExtraLib
+
+void printCredits();
+void printDebug(Categories *DATA);
+
+#endif