+++ app/lib/MessageLib.h
@@ -3,12 +3,14 @@
 
 void fetchMessages(Categories *DATA, unsigned int numCat);
 
-void fillMessageLog(Categories *DATA, unsigned int numCat, char *messageTitle);
+void fillMessageLog(Categories *DATA, unsigned int numCat);
 
 void printMesList(Categories *DATA, unsigned int numCat);
 
 void askForTextInput(char *instruction, char **unallocatedStr, size_t maxLength);
 
 void createMessage(Categories *DATA, unsigned int numCat);
+
+void deleteMessage(Categories *DATA, unsigned numCat, unsigned messageToDeleteNumber);
 
 #endif // MessageLib